suggested changes, add qmake to buildsystems
