+++ lib/buildsystems/qmake.rb
@@ -1,0 +1,13 @@
+require 'package'
+
+class Qmake < Package
+
+  def self.build
+    system 'qmake'
+    system 'make'
+  end
+
+  def self.install
+    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
+  end
+end
