+++ packages/qt5_location.rb
@@ -1,6 +1,6 @@
-require 'package'
+require 'buildsystems/qmake'
 
-class Qt5_location < Package
+class Qt5_location < Qmake
   description 'Qt Location and Positioning'
   homepage 'https://www.qt.io/'
   version '5.15.11-48a17e8'
@@ -27,13 +27,4 @@ class Qt5_location < Package
   depends_on 'qt5_base' # R
   depends_on 'qt5_declarative' # R
   depends_on 'zlibpkg' # R
-
-  def self.build
-    system 'qmake'
-    system 'make'
-  end
-
-  def self.install
-    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
-  end
 end
