+++ packages/qt5_multimedia.rb
@@ -1,6 +1,6 @@
-require 'package'
+require 'buildsystems/qmake'
 
-class Qt5_multimedia < Package
+class Qt5_multimedia < Qmake
   description 'Qt Multimedia'
   homepage 'https://www.qt.io/'
   version '5.15.11-f587b18'
@@ -30,13 +30,4 @@ class Qt5_multimedia < Package
   depends_on 'pulseaudio' # R
   depends_on 'qt5_base' # R
   depends_on 'qt5_declarative' # R
-
-  def self.build
-    system 'qmake'
-    system 'make'
-  end
-
-  def self.install
-    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
-  end
 end
