+++ packages/qt5_script.rb
@@ -1,6 +1,6 @@
-require 'package'
+require 'buildsystems/qmake'
 
-class Qt5_script < Package
+class Qt5_script < Qmake
   description 'Qt Script Tools'
   homepage 'https://www.qt.io/'
   version '5.15.15'
@@ -24,13 +24,4 @@ class Qt5_script < Package
   depends_on 'glibc' # R
   depends_on 'libglvnd' # R
   depends_on 'qt5_base' # R
-
-  def self.build
-    system 'qmake'
-    system 'make'
-  end
-
-  def self.install
-    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
-  end
 end
