+++ packages/qt5_serialport.rb
@@ -1,6 +1,6 @@
-require 'package'
+require 'buildsystems/qmake'
 
-class Qt5_serialport < Package
+class Qt5_serialport < Qmake
   description 'Qt Serial Port'
   homepage 'https://www.qt.io/'
   version '5.15.11-3380465'
@@ -24,13 +24,4 @@ class Qt5_serialport < Package
   depends_on 'gcc_lib' # R
   depends_on 'glibc' # R
   depends_on 'qt5_base' # R
-
-  def self.build
-    system 'qmake'
-    system 'make'
-  end
-
-  def self.install
-    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
-  end
 end
