+++ packages/qt5_webglplugin.rb
@@ -1,6 +1,6 @@
-require 'package'
+require 'buildsystems/qmake'
 
-class Qt5_webglplugin < Package
+class Qt5_webglplugin < Qmake
   description 'Qt WebGL Plugin'
   homepage 'https://www.qt.io/'
   version '5.15.11-ddcff45'
@@ -31,13 +31,4 @@ class Qt5_webglplugin < Package
   depends_on 'qt5_declarative' # R
   depends_on 'qt5_websockets' # R
   depends_on 'zlibpkg' # R
-
-  def self.build
-    system 'qmake'
-    system 'make'
-  end
-
-  def self.install
-    system "make INSTALL_ROOT=#{CREW_DEST_DIR} install"
-  end
 end
